backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow
numpy
pandas
google-generativeai
pydantic
python-dotenv
google-api-python-client
